split response by restful/graphql
