+++ packages/mock-server/server.ts
@@ -1,4 +1,5 @@
 import { setupServer } from "msw/node";
-import { requestHandlers } from "./request-handlers";
+import { requestHandlers as restfulHandlers } from "./handlers-restful/request-handlers";
+import { requestHandlers as graphQLHandlers } from "./handlers-graphql/request-handlers";
 
-export const server = setupServer(...requestHandlers);
+export const server = setupServer(...restfulHandlers, ...graphQLHandlers);
